requirements.txt
tk
sounddevice==0.4.7
numpy==1.26.4
pyttsx3==2.90
faster-whisper==1.0.3
python-dotenv==1.0.1
openai==1.40.6
